Create an umbrella build for NCEPLIBS (#76)
Utilize the ExternalProject functionality of CMake to create an umbrella build of NCEPLIBS using their develop branches.

CRTM is enabled as a default option.

EMC_post is enabled as a default option.
Utilities such as grib_util, prod_util, UFS_UTILS are not made a part of this umbrella build.
Executables from applications such as EMC_post is not a part of this umbrella build.

On a generic system with a compiler, MPI and NetCDF, the build instructions are:

 Load your modules or set environment variables
 e.g. 
 module load gnu
 module load openmpi
 module load netcdf

export CC=$(which mpicc)
export FC=$(which mpif90)
export CXX=$(which mpicxx)

git clone https://github.com/noaa-emc/nceplibs -b feature/umbrella
cd nceplibs
mkdir -p build && cd build
cmake -DCMAKE_INSTALL_PREFIX=../nceplibs-install -DDEPLOY=ON ..
make -j4 
make deploy
```

`-DDEPLOY=ON` will create a file deploy_modulefiles.sh and a target deploy. The target executes deploy_modulefiles.sh.

 Default values in deploy_modulefiles.sh are filled with user given options during configuration.

Choice of library branch/tag and version is provided via a table in file COMPONENTS.

All libraries, associated headers and fortran modules will be installed under ./nceplibs-install with the install tree look like:


./nceplibs-install
├── bacio
│   └── bacio-x.y.z
│       ├── include_4
│       ├── include_8
│       └── lib
├── nemsio
│   └── nemsio-develop
│       ├── bin
│       ├── include
│       └── lib
└── w3nco
    └── w3nco-feature-tryme
        ├── include_4
        ├── include_8
        ├── include_d
        └── lib

The modulefiles are installed under <CMAKE_INSTALL_PREFIX>/modules. This can be changed in deploy_modulefiles.sh or executing that script with a command-line argument.
The default tree structure of the module files is:


./nceplibs-install/modules
├── bacio
│   └── bacio-x.y.z.lua
├── nceppost
│   └── nceppost-feature-tryme.lua
└── w3nco
    └── w3nco-develop.lua

This module structure is reflective of the install tree structure.